Migrates PublishedContentExtensions.cs, PublishedElementExtensions.cs and PublishedPropertyExtension.cs
